mediverify/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
